Added lots of text, moved Extensor to utils, and tweaked the solution inputs parameters.
